Add files via upload
Assignment_1: The purpose of this MATLAB program is to create a comparative table of car names with their corresponding Carbon Dioxide emissions per year. The standard given for all cars was that 1 gallon of gas burned emits 19.4 lbs. of CO2 and that each car traveled 12,000 miles per year. 

Assignment_2: This MATLAB program determines the optimal trajectory (or angle (theta)) for maximizing the range of an object, launched at three given initial velocities of 50 m/s, 100 m/s and 200 m/s, respectively.

The purpose of this exercise is to exemplify comprehensive MATLAB skills including but not limited to plotting, and animated plotting of data.